requirements.txt
django
django-environ
django-cors-headers
djangorestframework

python-dotenv
langchain
langchain-openai
langchain-google-genai
Pillow


uvicorn
gunicorn